ui: add autocomplete for supervisors
* Creates LiteratureAuthorsField to share autocompleted affiliation and
full_name for authors and supersivors (of literature submission)

* INSPIR-2405
